requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
asgiref==3.5.2
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
bokeh==2.2.3
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==1.0.0
Django==3.2.14
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.21.3
entrypoints==0.3
flake8==3.8.4
Flask==1.1.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.33.2
gunicorn==20.1.0
h5py==3.0.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
llvmlite==0.34.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
mxnet==1.7.0.post1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
notebook==6.1.4
npm==0.1.1
numba==0.51.2
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
optional-django==0.1.0
packaging==21.3
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psycopg2-binary==2.9.3
ptyprocess==0.6.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pygame==2.0.0
Pygments==2.7.2
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyrsistent==0.17.3
pyspark==3.0.1
python-dateutil==2.8.1
pytz==2022.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
statsmodels==0.12.1
sudo==1.0.0
sympy==1.6.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.2
torch==1.3.1+cpu
torchvision==0.4.2+cpu
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==6.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
